comfyui-nodes/requirements.txt
aiohttp
numpy
Pillow
scipy
